Share primitive binary operand decoding
Move primitive array and constant decoding out of numeric execution so comparison does not depend on a numeric implementation detail. Numeric and comparison keep the same input classification and execution behavior.
