+++ vortex-array/src/scalar_fn/fns/binary/mod.rs
@@ -43,6 +43,7 @@ mod compare;
 pub use compare::*;
 mod numeric;
 pub(crate) use numeric::*;
+mod primitive_operand;
 
 use crate::scalar::NumericOperator;
 use crate::scalar::Scalar;
